source_code/ai-service/requirements.txt
SQLAlchemy==2.0.41
PyMySQL==1.1.1
python-dotenv==1.0.1
cryptography==45.0.3